Wrap up today's project
I made dramatic progress on the in-game menu system. Cardboard VR games
only have one button, and this makes it hard to give the player a
diverse set controls. To solve this problem I created a user-friendly
in-game menu. To move the character the player simply has to look at a
button.